// back/src/reserve/dto/select-reserve.dto.ts
export class SelectReserveDto {
    CORESE: number;
    COARTI: string;
    QTERESE: string;
    MAGASIN: string;
    IMPUTABLE: string;
    NAF: string;
    GACLEUNIK: number;
    NACLEUNIK: number;
    CODEEMP: string;
    DAT: string;
    I1CLEUNIK: number;
    NAFORIGINE: string;
    EMPARTQTE: string;
    ARTNAF: string;
    NAFCOARTI: string;
    LOCLEUNIK: string;
    COARTIMAGASIN: string;
}