// back/src/stock/dto/create-stock.dto.ts
export class CreateStockDto {
    STCLEUNIK: number;
    LOCLEUNIK: number;
    RBCLEUNIK: number;
    COARTI: string;
    DESA1: string;
    DESA2: string;
    COFA: string;
    ENTSO: string;
    QTE: string;
    MAGASIN: number;
    MAGORIGINE: number;
    CORESE: number;
    DAT: string;
    HEURE: string;
    CODEEMP: string;
    NAF: number;
    GACLEUNIK: number;
    NACLEUNIK: number;
    NAFORIGINE: number;
    STORIGINE: string;
    IMPUTABLE: string;
    MONTANT: string;
    DIVERS: string;
    LIBRENUME1: number;
    LIBRENUME2: number;
    DIM1: string;
    DIM2: string;
    OBS: string;
    TYPEALEA: number;
    SPECIFIQUE: number;
    I1CLEUNIK: number;
    ABCLEUNIK: number;
    SVCLEUNIK: number;
    DMCLEUNIK: number;
    NAFCOARTI: string;
    COARTIDATEHEURE: string;
}