// back/src/recbl/dto/select-recbl.dto.ts
export class SelectRecblDto {
    RBCLEUNIK: number;
  
    NBLFOURN: string;
  
    DATEREC: string;
  
    COFOU: string;
  
    NAF: number;
  
    TYPEBC: number;
  
    LBCLEUNIK: string;
  
    DESA1: string;
  
    QTE: string;
  
    Unite: string;
  
    QteRecep: string;
  
    UniteRecep: number;
  
    CoefRecep: number;
  
    QteRetour: number;
  
    CoefUnite: string;
  
    UnitePrix: number;
  
    CoefPrix: number;
  
    COBC: string;
  
    DATESAISIE: string;
  
    FLAG: number;
  
    QTESTOCK: string;
  
    CODEEMP: string;
  
    KGMT: string;
  
    OBS: string;
  
    BLCHIFFRE: string;
  
    FOUDATEBL: string;
}